Remove primer and simplify benchmark crate
Drop the primer workflow, script, composite action, and supporting
references. The primer was flaky and not pulling its weight.

While here, flatten the benchmark crate that was originally built to
support many real-world projects. With only one benchmark project, the
RealWorldProject/InstalledProject abstraction was overhead, so collapse
the three modules into a single lib.rs that hard-codes the karva-benchmark-1
settings, drop the unused criterion shim, swap ProjectMetadata::discover
for ProjectMetadata::new since the metadata is constructed directly, and
remove the dynamic per-bench-file matrix from CI in favor of running the
single benchmark by name.
