--- .github/actions/post-primer-comment/action.yml
@@ -1,53 +1,0 @@
-name: "Post primer comment"
-description: "Post or update the primer results PR comment"
-
-inputs:
-  pr-number:
-    description: "Pull request number to comment on"
-    required: true
-  comment-path:
-    description: "Path to the markdown file containing the comment body"
-    required: false
-    default: "primer-comment.md"
-
-runs:
-  using: "composite"
-  steps:
-    - uses: actions/github-script@f28e40c7f34bde8b3046d885e986cb6290c5673b # v7.1.0
-      env:
-        PRIMER_PR_NUMBER: ${{ inputs.pr-number }}
-        PRIMER_COMMENT_PATH: ${{ inputs.comment-path }}
-      with:
-        script: |
-          const fs = require('fs');
-          const path = process.env.PRIMER_COMMENT_PATH;
-          if (!fs.existsSync(path)) {
-            core.warning(`${path} not found — primer may have crashed before writing output.`);
-            return;
-          }
-          const body = fs.readFileSync(path, 'utf8');
-          const marker = '<!-- primer-results -->';
-          const issue_number = parseInt(process.env.PRIMER_PR_NUMBER, 10);
-
-          const { data: comments } = await github.rest.issues.listComments({
-            owner: context.repo.owner,
-            repo: context.repo.repo,
-            issue_number,
-          });
-
-          const existing = comments.find(c => c.body.includes(marker));
-          if (existing) {
-            await github.rest.issues.updateComment({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              comment_id: existing.id,
-              body,
-            });
-          } else {
-            await github.rest.issues.createComment({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              issue_number,
-              body,
-            });
-          }
