--- .github/workflows/primer-comment.yml
@@ -1,47 +1,0 @@
-name: "Primer Comment"
-
-on:
-  workflow_run:
-    workflows: ["Primer"]
-    types:
-      - completed
-
-permissions: {}
-
-jobs:
-  comment:
-    name: "Post primer comment"
-
-    if: github.event.workflow_run.event == 'pull_request'
-
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-
-    permissions:
-      pull-requests: write
-
-    steps:
-      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4.3.1
-        with:
-          persist-credentials: false
-          sparse-checkout: .github/actions/post-primer-comment
-
-      - name: "Download primer results"
-        id: download
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4.3.0
-        continue-on-error: true
-        with:
-          name: primer-result
-          run-id: ${{ github.event.workflow_run.id }}
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: "Read PR number"
-        id: pr
-        if: steps.download.outcome == 'success'
-        run: echo "number=$(cat pr-number.txt)" >> "$GITHUB_OUTPUT"
-
-      - name: "Post PR comment"
-        if: steps.download.outcome == 'success'
-        uses: ./.github/actions/post-primer-comment
-        with:
-          pr-number: ${{ steps.pr.outputs.number }}
