--- .github/workflows/primer.yml
@@ -1,103 +1,0 @@
-# Runs the primer script on each PR and posts the results as a comment.
-#
-# Posting the comment is split based on where the PR comes from. For PRs
-# from branches in this repo, the GITHUB_TOKEN already has write access,
-# so we post the comment directly from this workflow. For PRs from forks,
-# GitHub forces the token to read-only regardless of the permissions block,
-# so we instead upload the results as an artifact and let primer-comment.yml
-# (triggered by workflow_run, which always runs in the base repo context
-# with full permissions) post the comment. Both paths share the same
-# comment-posting logic via the post-primer-comment composite action.
-
-name: Primer
-
-on:
-  pull_request:
-  workflow_dispatch:
-
-permissions: {}
-
-jobs:
-  primer:
-    name: "primer"
-
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-
-    permissions:
-      # Granted for same-repo PRs so we can post the comment directly.
-      # Forks get this downgraded to read-only by GitHub regardless, and
-      # fall back to the primer-comment.yml workflow_run path.
-      pull-requests: write
-
-    steps:
-      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4.3.1
-        with:
-          fetch-depth: 0
-          persist-credentials: false
-
-      - uses: Swatinem/rust-cache@e18b497796c12c097a38f9edb9d0641fb99eee32 # v2
-
-      - name: "Install Rust toolchain"
-        run: rustup show
-
-      - uses: astral-sh/setup-uv@d4b2f3b6ecc6e67c4457f6d3e41ec42d3d0fcb86 # v5.4.2
-        with:
-          python-version: "3.13"
-
-      - name: "Cache primer projects"
-        uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4.3.0
-        with:
-          path: target/primer_projects
-          key: primer-${{ hashFiles('scripts/primer.py') }}
-          restore-keys: |
-            primer-
-
-      - name: "Build branch wheel"
-        run: uvx maturin build --out target/wheels-pr/
-
-      - name: "Build main baseline wheel"
-        if: github.event_name == 'pull_request'
-        run: |
-          git worktree add /tmp/karva-main origin/main
-          cd /tmp/karva-main
-          uvx maturin build --out "$GITHUB_WORKSPACE/target/wheels-main/"
-          cd "$GITHUB_WORKSPACE"
-          git worktree remove /tmp/karva-main
-
-      - name: "Run primer"
-        run: |
-          uv run --script scripts/primer.py -v \
-            --wheel "$(ls target/wheels-pr/*.whl)" \
-            --baseline-wheel "$(ls target/wheels-main/*.whl)" \
-            --markdown-output primer-comment.md
-
-      - name: "Post PR comment"
-        if: >
-          !cancelled()
-          && github.event_name == 'pull_request'
-          && github.event.pull_request.head.repo.full_name == github.repository
-        uses: ./.github/actions/post-primer-comment
-        with:
-          pr-number: ${{ github.event.number }}
-
-      - name: "Save PR number"
-        if: >
-          !cancelled()
-          && github.event_name == 'pull_request'
-          && github.event.pull_request.head.repo.full_name != github.repository
-        env:
-          PR_NUMBER: ${{ github.event.number }}
-        run: echo "$PR_NUMBER" > pr-number.txt
-
-      - name: "Upload primer results"
-        if: >
-          !cancelled()
-          && github.event_name == 'pull_request'
-          && github.event.pull_request.head.repo.full_name != github.repository
-        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
-        with:
-          name: primer-result
-          path: |
-            primer-comment.md
-            pr-number.txt
