+++ CONTRIBUTING.md
@@ -58,7 +58,7 @@ Karva uses a **main-process + worker-subprocess** execution model. When you run 
 
 **Dev / Testing:**
 
-- `karva_benchmark` — Wall-time benchmarks using divan, including real-world project definitions.
+- `karva_benchmark` — Wall-time benchmark for karva, runs `karva test` against a pinned snapshot of `karva-benchmark-1`.
 
 ### Key Design Decisions
 
@@ -108,19 +108,6 @@ Or simply, with just, run:
 
 ```bash
 just test
-```
-
-### Primer (real-world compatibility testing)
-
-The primer builds a fresh karva wheel, clones a curated list of popular Python
-projects, installs their dependencies, and runs karva against each one to
-validate end-to-end compatibility.
-
-```bash
-uv run --script scripts/primer.py            # build wheel + run all projects
-uv run --script scripts/primer.py -v         # stream full karva output
-uv run --script scripts/primer.py -p httpx   # run a single project
-uv run --script scripts/primer.py --help     # show all options
 ```
 
 ### Documentation
