+++ crates/karva_benchmark/Cargo.toml
@@ -14,7 +14,6 @@ license = { workspace = true }
 [dependencies]
 anyhow = { workspace = true }
 camino = { workspace = true }
-codspeed-criterion-compat = { workspace = true }
 divan = { workspace = true }
 karva_cli = { workspace = true }
 karva_logging = { workspace = true }
@@ -25,7 +24,6 @@ karva_static = { workspace = true }
 ruff_python_ast = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-tracing = { workspace = true }
 
 [[bench]]
 name = "karva_benchmark_walltime"
