--- crates/karva_benchmark/src/criterion.rs
@@ -1,1 +1,0 @@
-pub use codspeed_criterion_compat::*;
